Commit Information: 1. almost finished

+++ app/src/main/java/dev/bogibek/nutritionxorazm/models/Advice.kt
@@ -10,5 +10,5 @@ data class AdviceResponse(
 
 data class Recommend(
     val name: String,
-    val ccal: Long,
+    val ccal: Int,
 )
